requirements.txt
flask==2.1.2
flask-sqlalchemy==2.5.1
flask-restful==0.3.9
python-dotenv==0.20.0
flask-marshmallow==0.14.0
sqlalchemy-filters==0.12.0
SQLAlchemy~=1.4.36
mysqlclient==1.4.6
marshmallow~=3.16.0
